refactor: offload editor functionality to components
