+++ src/data/document_stats.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Default, Clone, Copy)]
+pub struct DocumentStats {
+    pub num_lines: i32,
+    pub num_chars: i32,
+    pub num_spaces: i32,
+    pub num_words: i32,
+}
