+++ src/data/mod.rs
@@ -1,5 +1,7 @@
 mod document_data;
+mod document_stats;
 mod folder_data;
 
 pub use document_data::DocumentObject;
+pub use document_stats::DocumentStats;
 pub use folder_data::FolderObject;
